+++ src/core/components/controls/cpu/cpufreq.cpp
@@ -5,16 +5,19 @@
 
 #include "core/icommandqueue.h"
 #include <algorithm>
+#include <cassert>
 #include <utility>
 
 CPUFreq::CPUFreq(std::vector<std::string> &&scalingGovernors,
                  std::string const &defaultGovernor,
                  std::vector<std::unique_ptr<IDataSource<std::string>>>
-                     &&scalingGovernorDataSources) noexcept
+                     &&scalingGovernorDataSources,
+                 std::unique_ptr<IEPPHandler> &&eppHandler) noexcept
 : Control(true)
 , id_(CPUFreq::ItemID)
 , scalingGovernors_(std::move(scalingGovernors))
 , scalingGovernorDataSources_(std::move(scalingGovernorDataSources))
+, eppHandler_(std::move(eppHandler))
 {
   scalingGovernor(defaultGovernor);
   if (scalingGovernor_.empty())
@@ -40,15 +43,23 @@ std::string const &CPUFreq::ID() const
 
 void CPUFreq::importControl(IControl::Importer &i)
 {
-  auto &cpuFreqImporter = dynamic_cast<CPUFreq::Importer &>(i);
-  scalingGovernor(cpuFreqImporter.provideCPUFreqScalingGovernor());
+  auto &importer = dynamic_cast<CPUFreq::Importer &>(i);
+  scalingGovernor(importer.provideCPUFreqScalingGovernor());
+
+  if (eppHandler_) {
+    auto hint = importer.provideCPUFreqEPPHint();
+    assert(hint.has_value());
+    eppHandler_->hint(*hint);
+  }
 }
 
 void CPUFreq::exportControl(IControl::Exporter &e) const
 {
-  auto &cpuFreqExporter = dynamic_cast<CPUFreq::Exporter &>(e);
-  cpuFreqExporter.takeCPUFreqScalingGovernors(scalingGovernors());
-  cpuFreqExporter.takeCPUFreqScalingGovernor(scalingGovernor());
+  auto &exporter = dynamic_cast<CPUFreq::Exporter &>(e);
+  exporter.takeCPUFreqScalingGovernors(scalingGovernors());
+  exporter.takeCPUFreqEPPHints(eppHints());
+  exporter.takeCPUFreqScalingGovernor(scalingGovernor());
+  exporter.takeCPUFreqEPPHint(eppHint());
 }
 
 void CPUFreq::cleanControl(ICommandQueue &)
@@ -62,6 +73,9 @@ void CPUFreq::syncControl(ICommandQueue &ctlCmds)
       if (dataSourceEntry_ != scalingGovernor())
         ctlCmds.add({scalingGovernorDataSource->source(), scalingGovernor()});
     }
+
+  if (eppHandler_)
+    eppHandler_->sync(ctlCmds);
 }
 
 std::string const &CPUFreq::scalingGovernor() const
@@ -84,4 +98,14 @@ void CPUFreq::scalingGovernor(std::string const &scalingGovernor)
 std::vector<std::string> const &CPUFreq::scalingGovernors() const
 {
   return scalingGovernors_;
+}
+
+std::optional<std::string> CPUFreq::eppHint() const
+{
+  return eppHandler_ ? std::make_optional(eppHandler_->hint()) : std::nullopt;
+}
+
+std::optional<std::vector<std::string>> CPUFreq::eppHints() const
+{
+  return eppHandler_ ? std::make_optional(eppHandler_->hints()) : std::nullopt;
 }