+++ src/core/components/controls/cpu/cpufreq.h
@@ -4,8 +4,10 @@
 #pragma once
 
 #include "core/components/controls/control.h"
+#include "core/components/controls/cpu/handlers/iepphandler.h"
 #include "core/idatasource.h"
 #include <memory>
+#include <optional>
 #include <string>
 #include <string_view>
 #include <vector>
@@ -19,6 +21,7 @@ class CPUFreq : public Control
   {
    public:
     virtual std::string const &provideCPUFreqScalingGovernor() const = 0;
+    virtual std::optional<std::string> const &provideCPUFreqEPPHint() const = 0;
   };
 
   class Exporter : public IControl::Exporter
@@ -27,12 +30,16 @@ class CPUFreq : public Control
     virtual void takeCPUFreqScalingGovernor(std::string const &governor) = 0;
     virtual void
     takeCPUFreqScalingGovernors(std::vector<std::string> const &governors) = 0;
+    virtual void takeCPUFreqEPPHint(std::optional<std::string> const &hint) = 0;
+    virtual void
+    takeCPUFreqEPPHints(std::optional<std::vector<std::string>> const &hints) = 0;
   };
 
   CPUFreq(std::vector<std::string> &&scalingGovernors,
           std::string const &defaultGovernor,
           std::vector<std::unique_ptr<IDataSource<std::string>>>
-              &&scalingGovernorDataSources) noexcept;
+              &&scalingGovernorDataSources,
+          std::unique_ptr<IEPPHandler> &&eppHandler = nullptr) noexcept;
 
   void preInit(ICommandQueue &ctlCmds) final override;
   void postInit(ICommandQueue &ctlCmds) final override;
@@ -49,14 +56,17 @@ class CPUFreq : public Control
 
   std::string const &scalingGovernor() const;
   void scalingGovernor(std::string const &governor);
-
   std::vector<std::string> const &scalingGovernors() const;
+
+  std::optional<std::string> eppHint() const;
+  std::optional<std::vector<std::string>> eppHints() const;
 
  private:
   std::string const id_;
   std::vector<std::string> const scalingGovernors_;
   std::vector<std::unique_ptr<IDataSource<std::string>>> const
       scalingGovernorDataSources_;
+  std::unique_ptr<IEPPHandler> eppHandler_;
 
   std::string scalingGovernor_;
   std::string dataSourceEntry_;