+++ src/core/components/controls/cpu/cpufreqprofilepart.h
@@ -5,6 +5,7 @@
 
 #include "core/profilepart.h"
 #include "cpufreq.h"
+#include <optional>
 #include <string>
 #include <vector>
 
@@ -17,12 +18,14 @@ class CPUFreqProfilePart final
   {
    public:
     virtual std::string const &provideCPUFreqScalingGovernor() const = 0;
+    virtual std::optional<std::string> const &provideCPUFreqEPPHint() const = 0;
   };
 
   class Exporter : public IProfilePart::Exporter
   {
    public:
     virtual void takeCPUFreqScalingGovernor(std::string const &governor) = 0;
+    virtual void takeCPUFreqEPPHint(std::optional<std::string> const &hint) = 0;
   };
 
   CPUFreqProfilePart() noexcept;
@@ -38,6 +41,7 @@ class CPUFreqProfilePart final
 
   bool provideActive() const override;
   std::string const &provideCPUFreqScalingGovernor() const override;
+  std::optional<std::string> const &provideCPUFreqEPPHint() const override;
 
  protected:
   void importProfilePart(IProfilePart::Importer &i) override;
@@ -46,12 +50,16 @@ class CPUFreqProfilePart final
 
  private:
   void governor(std::string const &governor);
+  void eppHint(std::optional<std::string> const &hint);
 
   class Initializer;
 
   std::string const id_;
   std::string governor_;
   std::vector<std::string> governors_;
+
+  std::optional<std::string> eppHint_;
+  std::optional<std::vector<std::string>> eppHints_;
 
   static bool const registered_;
 };